+++ arch/include/arch_vcpu.h
@@ -69,7 +69,7 @@ u32 arch_vcpu_irq_priority(struct vmm_vcpu *vcpu, u32 irq_no);
  *  NOTE: This function is usually useful to architectures having
  *  hardware virtualization support.
  */
-int arch_vcpu_irq_assert(struct vmm_vcpu *vcpu, u32 irq_no, u32 reason);
+int arch_vcpu_irq_assert(struct vmm_vcpu *vcpu, u32 irq_no, u64 reason);
 
 /** Execute VCPU interrupt 
  *  NOTE: This functions is always called in context of the VCPU (i.e.
@@ -79,13 +79,13 @@ int arch_vcpu_irq_assert(struct vmm_vcpu *vcpu, u32 irq_no, u32 reason);
  */
 int arch_vcpu_irq_execute(struct vmm_vcpu *vcpu, 
 			  arch_regs_t *regs,
-			  u32 irq_no, u32 reason);
+			  u32 irq_no, u64 reason);
 
 /** Deassert VCPU interrupt 
  *  NOTE: This functions is called asynchronusly in any context.
  *  NOTE: This function is usually useful to architectures having
  *  hardware virtualization support.
  */
-int arch_vcpu_irq_deassert(struct vmm_vcpu * vcpu, u32 irq_no, u32 reason);
+int arch_vcpu_irq_deassert(struct vmm_vcpu *vcpu, u32 irq_no, u64 reason);
 
 #endif
